+++ emails/validators.py
@@ -1,7 +1,6 @@
 """Field validators for emails models."""
 
 import re
-from typing import Any
 
 from django.contrib.auth.models import User
 
@@ -11,7 +10,6 @@ from .apps import BadWords, emails_config
 from .exceptions import (
     AccountIsInactiveException,
     AccountIsPausedException,
-    CannotMakeSubdomainException,
     DomainAddrFreeTierException,
     DomainAddrNeedSubdomainException,
     RelayAddrFreeTierLimitException,
@@ -21,11 +19,6 @@ from .exceptions import (
 #   can't start or end with a hyphen
 #   must be 1-63 lowercase alphanumeric characters and/or hyphens
 _re_valid_address = re.compile("^(?![-.])[a-z0-9-.]{1,63}(?<![-.])$")
-
-# A valid subdomain:
-#   can't start or end with a hyphen
-#   must be 1-63 alphanumeric characters and/or hyphens
-_re_valid_subdomain = re.compile("^(?!-)[a-z0-9-]{1,63}(?<!-)$")
 
 
 def badwords() -> BadWords:
@@ -104,29 +97,3 @@ def valid_address(address: str, domain: str, subdomain: str | None = None) -> bo
 def valid_address_pattern(address: str) -> bool:
     """Return if the local/user part of an address is valid."""
     return _re_valid_address.match(address) is not None
-
-
-def valid_available_subdomain(subdomain: Any) -> None:
-    """Raise CannotMakeSubdomainException if the subdomain fails a validation test."""
-    from .models import RegisteredSubdomain, hash_subdomain
-
-    if not subdomain:
-        raise CannotMakeSubdomainException("error-subdomain-cannot-be-empty-or-null")
-    subdomain = str(subdomain).lower()
-
-    # valid subdomains:
-    #   have to meet the rules for length and characters
-    valid = _re_valid_subdomain.match(subdomain) is not None
-    #   can't have "bad" words in them
-    bad_word = has_bad_words(subdomain)
-    #   can't have "blocked" words in them
-    blocked_word = is_blocklisted(subdomain)
-    #   can't be taken by someone else
-    taken = (
-        RegisteredSubdomain.objects.filter(
-            subdomain_hash=hash_subdomain(subdomain)
-        ).count()
-        > 0
-    )
-    if not valid or bad_word or blocked_word or taken:
-        raise CannotMakeSubdomainException("error-subdomain-not-available")
