+++ privaterelay/settings.py
@@ -458,10 +458,12 @@ SUBSCRIPTIONS_WITH_VPN: list[str] = config(
 MAX_ONBOARDING_AVAILABLE = config("MAX_ONBOARDING_AVAILABLE", 0, cast=int)
 MAX_ONBOARDING_FREE_AVAILABLE = config("MAX_ONBOARDING_FREE_AVAILABLE", 3, cast=int)
 
-MAX_ADDRESS_CREATION_PER_DAY = config("MAX_ADDRESS_CREATION_PER_DAY", 100, cast=int)
-MAX_REPLIES_PER_DAY = config("MAX_REPLIES_PER_DAY", 100, cast=int)
-MAX_FORWARDED_PER_DAY = config("MAX_FORWARDED_PER_DAY", 1000, cast=int)
-MAX_FORWARDED_EMAIL_SIZE_PER_DAY = config(
+MAX_ADDRESS_CREATION_PER_DAY: int = config(
+    "MAX_ADDRESS_CREATION_PER_DAY", 100, cast=int
+)
+MAX_REPLIES_PER_DAY: int = config("MAX_REPLIES_PER_DAY", 100, cast=int)
+MAX_FORWARDED_PER_DAY: int = config("MAX_FORWARDED_PER_DAY", 1000, cast=int)
+MAX_FORWARDED_EMAIL_SIZE_PER_DAY: int = config(
     "MAX_FORWARDED_EMAIL_SIZE_PER_DAY", 1_000_000_000, cast=int
 )
 PREMIUM_FEATURE_PAUSED_DAYS: int = config(
